added checklist renderer
