+++ src/types/icons.d.ts
@@ -2,4 +2,6 @@
 declare module '@codexteam/icons' {
   export const IconListBulleted: string;
   export const IconListNumbered: string;
+  // export const IconCheckList: string
+  export const IconCheck: string;
 }
