+++ src/types/itemMeta.ts
@@ -1,0 +1,6 @@
+export default interface ItemMeta {
+  /**
+   * State of the item
+   */
+  checked?: boolean;
+}
